// frontend/generated/models/SkillDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SkillDto
 */
export interface SkillDto {
    /**
     * 
     * @type {number}
     * @memberof SkillDto
     */
    skillId?: number;
    /**
     * 
     * @type {string}
     * @memberof SkillDto
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SkillDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof SkillDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof SkillDto
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the SkillDto interface.
 */
export function instanceOfSkillDto(value: object): value is SkillDto {
    return true;
}

export function SkillDtoFromJSON(json: any): SkillDto {
    return SkillDtoFromJSONTyped(json, false);
}

export function SkillDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SkillDto {
    if (json == null) {
        return json;
    }
    return {
        
        'skillId': json['skillId'] == null ? undefined : json['skillId'],
        'name': json['name'] == null ? undefined : json['name'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function SkillDtoToJSON(json: any): SkillDto {
    return SkillDtoToJSONTyped(json, false);
}

export function SkillDtoToJSONTyped(value?: SkillDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'skillId': value['skillId'],
        'name': value['name'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}